requirements.txt
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
joblib==1.1.0
kafka-python==2.0.2
nltk==3.7
packaging==21.3
pika==1.2.1
pyparsing==3.0.9
redis==4.3.3
regex==2022.6.2
rq==1.10.1
tqdm==4.64.0
Unidecode==1.3.4
wrapt==1.14.1
